add fitch
