+++ js/server.js
@@ -1,0 +1,12 @@
+// import {renderSimilarList} from './util.js';
+// // дз 10
+
+//   const formId = document.querySelector('#upload-select-image');
+
+//   fetch('https://25.javascript.pages.academy/kekstagram/data')
+
+//   .then((response) => response.json())
+//   .then((wizards) => {
+//     renderSimilarList(wizards);
+//   })
+
